remove module source fields from modtm_telemetry resource as code review suggested
