+++ internal/provider/telemetry_resource_test.go
@@ -19,11 +19,9 @@ import (
 	"time"
 
 	toxiproxy "github.com/Shopify/toxiproxy/v2/client"
-	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/prashantv/gostub"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 )
 
@@ -952,107 +950,4 @@ resource "modtm_telemetry" "test" {
 }
 `, endpointAssignment, enabledAssignment, sb.String(), modulePath)
 	return r
-}
-
-// createModulesJson creates a modules.json file with a reference to a standard module (kv) and
-// reference to a child module of the key vault module (keys) in the root module.
-func createModulesJson() error {
-	if err := os.MkdirAll(".terraform/modules", 0755); err != nil {
-		return err
-	}
-	f, err := os.Create(".terraform/modules/modules.json")
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-	_, err = f.WriteString(`{
-  "Modules": [
-    {
-      "Key": "",
-      "Source": "",
-      "Dir": "."
-    },
-    {
-      "Key": "keys",
-      "Source": "registry.terraform.io/Azure/avm-res-keyvault-vault/azurerm//modules/key",
-      "Version": "0.6.1",
-      "Dir": ".terraform/modules/keys/modules/key"
-    },
-    {
-      "Key": "kv",
-      "Source": "registry.terraform.io/Azure/avm-res-keyvault-vault/azurerm",
-      "Version": "0.6.1",
-      "Dir": ".terraform/modules/kv"
-    },
-    {
-      "Key": "kv.keys",
-      "Source": "./modules/key",
-      "Dir": ".terraform/modules/kv/modules/key"
-    },
-    {
-      "Key": "kv.secrets",
-      "Source": "./modules/secret",
-      "Dir": ".terraform/modules/kv/modules/secret"
-    }
-  ]
-}
-`)
-	return err
-}
-
-func TestUpdateModuleSourceAndVersion(t *testing.T) {
-	require.NoError(t, createModulesJson())
-
-	tests := []struct {
-		name           string
-		modulePath     string
-		expectedModule TelemetryResourceModel
-	}{
-		{
-			name:       "Valid module path",
-			modulePath: ".terraform/modules/kv",
-			expectedModule: TelemetryResourceModel{
-				ModulePath:    types.StringValue(".terraform/modules/kv"),
-				ModuleVersion: types.StringValue("0.6.1"),
-				ModuleSource:  types.StringValue("registry.terraform.io/Azure/avm-res-keyvault-vault/azurerm"),
-			},
-		},
-		{
-			name:       "Module not found",
-			modulePath: ".terraform/modules/nonexistent",
-			expectedModule: TelemetryResourceModel{
-				ModulePath: types.StringValue(".terraform/modules/nonexistent"),
-			},
-		},
-		{
-			name:       "Direct call to submodule",
-			modulePath: ".terraform/modules/keys/modules/key",
-			expectedModule: TelemetryResourceModel{
-				ModulePath:    types.StringValue(".terraform/modules/keys/modules/key"),
-				ModuleVersion: types.StringValue("0.6.1"),
-				ModuleSource:  types.StringValue("registry.terraform.io/Azure/avm-res-keyvault-vault/azurerm//modules/key"),
-			},
-		},
-		{
-			name:       "Nested module",
-			modulePath: ".terraform/modules/kv/modules/key",
-			expectedModule: TelemetryResourceModel{
-				ModulePath:    types.StringValue(".terraform/modules/kv/modules/key"),
-				ModuleVersion: types.StringValue(""),
-				ModuleSource:  types.StringValue("./modules/key"),
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			data := &TelemetryResourceModel{
-				ModulePath: types.StringValue(tt.modulePath),
-			}
-
-			result := withModuleSourceAndVersion(data)
-
-			assert.Equal(t, tt.expectedModule, *result)
-		})
-	}
 }
